⚡ Dashboard refactor
